pkvm: x86: add vmexit handler for host vm
This patch implements the basic set of vmexit handler to be able to run
pkvm_host_vm, including hadlers for vmexit reason of cpuid, cr, msr and
xsetbv accesses.

This patch also defines a pkvm_main function which do the loop handling
of vmexit_handlers based on above different vmexit reason, and
__pkvm_vmx_vcpu_run at the tail of loop to trigger vmlaunch/vmresume.

The new dir "hyp" is created to hold such hypervisor runtime
routines, which should be running in root-mode.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>
Reviewed-by: Eddie Dong <eddie.dong@intel.com>