pkvm: x86: initialize IOMMU cap before deprivilege
pkvm is a post launch hypervisor which runs after the IOMMU driver. So
the DMAR table is already parsed by IOMMU driver, so the DMAR MMIO
address and size can be retrieved through the intel_iommu data
structure. These information will be used to do the IOMMU initialization
in the later patches.

The IOMMU in pkvm will leverage the VT-D nested translation for better
performance with using the first-level translation from host VM and the
second-level translation from EPT. As EPT will be used as a second-level
page table in IOMMU, need to pick up a both supported page table level
and page size by VT-D and VMX.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>