pkvm: x86: implement do_donate() helper for donating memory
Transferring ownership information of a memory region from one component
to another can be achieved using a "donate" operation, which results in
the previous owner losing access to the underlying pages entirely, and
current owner will control the pages.

Introduce a do_donate() helper for safely donating a memory region from
one component to another. Currently, only host_to_hyp donating is
implemented, but the code is easily extended to handle other
combinations and the permission checks for each component are reuseable.

Signed-off-by: Shaoqin Huang <shaoqin.huang@intel.com>