pkvm: x86: introduce vept to record guest EPT information
KVM-high manages a guest EPT for its nested VM, which shall be shadowed
to shadow EPT in pkvm hypervisor.

This patch introduces a per vCPU pgtable(vept) to save the guest EPT
related information from KVM-high when vmwrite(EPTP).

Each time when a vCPU vmwrites to EPTP, pkvm will setup the vept
according to the new value or just do nothing if it is the same
with the last one.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>