pkvm: x86: add support for pkvm to handle the nested EPT violation
Shadow EPT will be used as EPTP for nested VMs so when EPT violation
happened, let pkvm to do the first handling. pkvm can walk the guest
vEPT to see if the L2 GPA already has a proper mapping in vEPT or not.
If already has one, then pkvm can map in the shadow EPT - now just do
same mapping as vEPT.

For the case that vEPT is not present or vEPT contains a misconfig
mapping, then pkvm can directly back to the KVM-high for proper handling.
Another case is that if the EPT violation vmexit has pending event to be
injected, pkvm also needs back to the KVM-high for event injection.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>