pkvm: x86: use page state API in shadow EPT for normal VM
Add map_leaf & free_leaf override helper functions for shadow EPT, and
use page state API in these helper functions to support shadow EPT
invalidation, destroy and EPT violation for normal VM.

When map a page for a normal VM in shadow EPT, use the share API to mark
this page is shared which is previously owned by the host VM but now is
shared between the host VM and the normal VM. And when invalidate or
destroy shadow EPT, mark this page as unshared which means owned by the
host VM again.

Under the state machine of page state transition, we do not support
multiple guest pages mapping to same host page, it's conflict with KSM,
so just disable it under pkvm Kconfig.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>