pkvm: x86: introduce refcount for shadow_vcpu
Similar to the refcount in shadow_vm, the refcount in shadow_vcpu
also protects the shadow_vcpu data structure from malicous teardown.

The refcount is initialized as 1 when init a shadow_vcpu. To get a valid
shadow_vcpu pointer, needs to use get_shadow_vcpu with its shadow vcpu
handle. And To release this shadow_vcpu pointer, needs to use
put_shadow_vcpu with this handle. So if a shadow vcpu is still being
used, then it cannot be teardown.

The get/put_shadow_vcpu will be used by the later patch to run a shadow
vcpu.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>