pkvm: x86: intercept the APIC_ID msr
Primary VM can change the local APIC_ID by writing APIC_ID MSR in x2apic
mode. This usually not happen but still need to intercept this. If
primary VM is changing to a different APIC_ID, then assert to forbidden
this as this will result in pkvm failed to kick.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>